+++ test/test_snapshot.py
@@ -162,6 +162,12 @@ snapshots = [
     Snapshot("top_level_reimports", ["top_level_reimports"]),
     Snapshot("type_checking_imports"),
     Snapshot("type_stub", min_version=(3, 10)),
+    Snapshot(
+        "visibility",
+        render_options={
+            "include_undocumented": False,
+        },
+    ),
 ]